# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  common.ResponseError:
    properties:
      error:
        type: string
    type: object
...
  common.ResponseNormal:
    properties:
      data: {}
      email:
        type: string
      message:
        type: string
      paging: {}
      result:
        type: boolean
    type: object
...
  dtos.CreateActivityRequest:
    properties:
      description:
        type: string
      duration:
        type: integer
      level_id:
        type: string
      point_reward:
        type: integer
      repeat_days:
        items:
          type: string
        type: array
      title:
        type: string
      type:
        enum:
        - Activity
        - MiniGame
        - Challenge
        type: string
    required:
    - duration
    - level_id
    - point_reward
    - title
    - type
    type: object
...
  enum.ActivityType:
    enum:
    - Activity
    - MiniGame
    - Challenge
    type: string
    x-enum-varnames:
    - TypeActivity
    - TypeMiniGame
    - TypeChallenge
...
  models.Activity:
    properties:
      activity_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      duration:
        type: integer
      level_id:
        type: string
      point_reward:
        type: integer
      title:
        type: string
      type:
        $ref: '#/definitions/enum.ActivityType'
      updated_at:
        type: string
    required:
    - duration
    - point_reward
    - title
    type: object
...
host: 127.0.0.1:9000
info:
  contact: {}
  description: List APIs of Healthy Management Service
  termsOfService: http://swagger.io/terms/
  title: Healthy Service API Document
  version: "1.0"
paths:
...
  /expert/activities:
    get:
      consumes:
      - application/json
      description: Get activities for an expert
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Activities retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseNormal'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Activity'
                  type: array
              type: object
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/common.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.ResponseError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.ResponseError'
        "404":
          description: No activities found
          schema:
            $ref: '#/definitions/common.ResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get activities for an expert
      tags:
      - Activity
    post:
      consumes:
      - application/json
      description: Creates a new activity for a specified level, including repeat
        days. Requires valid JWT token.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Activity details to create
        in: body
        name: activity
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateActivityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Activity created successfully
          schema:
            $ref: '#/definitions/common.ResponseNormal'
        "400":
          description: invalid activity data, invalid level ID, invalid activity type,
            or invalid week day
          schema:
            $ref: '#/definitions/common.ResponseError'
        "401":
          description: token expired or invalid
          schema:
            $ref: '#/definitions/common.ResponseError'
        "404":
          description: level not found
          schema:
            $ref: '#/definitions/common.ResponseError'
        "500":
          description: failed to create activity
          schema:
            $ref: '#/definitions/common.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Create a new activity for a level
      tags:
      - Activity
...
schemes:
- http
- https
swagger: "2.0"